Add files via upload
Project for "Practice in Statistical Data Science"

Utilized statistical modeling techniques and an iterative approach to identify which factors impact the accuracy, or lack there of, in Zillow evaluations/estimates. In an ever-changing real estate market, the indicators of unpredictability can be explained by social trends, environmental factors, and other impactful trends in society. This analysis is complex and required lots of curiosity as well as trial and error.